feat: Add board post logic by userId

+++ workspace/apps/api/src/app/board/board.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException, BadRequestException } from '@nestjs/common';
+import { Injectable, NotFoundException, BadRequestException, UnauthorizedException } from '@nestjs/common';
 import { CreateCommentDto, CreatePostDto, UpdatePostDto, UpdateCommentDto } from '../common/dtos/post.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { UserService } from '../user/user.service';
@@ -26,7 +26,9 @@ export class BoardService {
             const response = await this.prisma.post.findUnique({
                 where: {
                     id: postId
-                }
+                },
+
+                include: { comments: true }
             });
 
             return response;
@@ -51,37 +53,49 @@ export class BoardService {
         }
     }
 
-    async updatePost(postId: string, updatePostDto: UpdatePostDto) {
+    async updatePost(postId: string, updatePostDto: UpdatePostDto, userId: string) {
         try {
-            const id = postId;
-            const updateData = {
-                slug: updatePostDto.slug,
-                title: updatePostDto.title,
-                body: updatePostDto.body,
-                published: updatePostDto.published
-            };
+            const postObject = await this.getPostById(postId);
+
+            if (!postObject) {
+                throw new BadRequestException('INVALID postId');
+            }
+
+            if (postObject.userId !== userId) {
+                throw new UnauthorizedException('INVALID userId');
+            }
 
             const response = await this.prisma.post.update({
                 where: {
-                    id: id
+                    id: postId
                 },
-                data: updateData
+                data: {
+                    ...updatePostDto
+                }
             });
 
             return response;
         }
-        catch(e) {
-            throw new BadRequestException(e);
+        catch(err) {
+            throw new BadRequestException(err);
         }
     }
 
-    async deletePost(postId: string) {
+    async deletePost(postId: string, userId: string) {
         try {
-            const id = postId;
+            const postObject = await this.getPostById(postId);
+
+            if (!postObject) {
+                throw new BadRequestException('INVALID postId');
+            }
+
+            if (postObject.userId !== userId) {
+                throw new UnauthorizedException('INVALID userId');
+            }
 
             const response = await this.prisma.post.delete({
                 where: {
-                    id: id
+                    id: postId
                 }
             });
